Provide backup for full outline of recipe website
For future reference I probably should have committed earlier in the
project as I made templates for each page that could be quickly
filled out with the information for each specific recipe.

This commit contains the main page and each linked recipe in various
stages of completion.

That being said, I'm pleased with all the progress I've made in just
a couple days of work on the odin project.